feat: add prototype framework
