+++ apps/prototypes/.gitignore
@@ -1,0 +1,2 @@
+.vercel
+.env*.local
