+++ apps/prototypes/next.config.mjs
@@ -1,0 +1,15 @@
+import { resolve } from 'path';
+
+const monorepoRoot = resolve(import.meta.dirname, '../..');
+
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  reactStrictMode: true,
+  poweredByHeader: false,
+  outputFileTracingRoot: monorepoRoot,
+  turbopack: {
+    root: monorepoRoot,
+  },
+};
+
+export default nextConfig;
