+++ apps/prototypes/src/app/kiloclaw-org-billing/prototype.json
@@ -1,0 +1,5 @@
+{
+  "title": "Organization KiloClaw billing UI",
+  "description": "Full-flow design review for organization-funded KiloClaw billing, including customer roles, admin support surfaces, lifecycle states, and launch readiness.",
+  "tags": ["Billing", "KiloClaw", "Organization"]
+}
