+++ apps/prototypes/src/app/layout.tsx
@@ -1,0 +1,43 @@
+import type { Metadata, Viewport } from 'next';
+import { Inter, JetBrains_Mono, Roboto_Mono } from 'next/font/google';
+import '@web/app/globals.css';
+
+const inter = Inter({
+  subsets: ['latin'],
+  display: 'swap',
+  variable: '--font-sans',
+});
+
+const mono = Roboto_Mono({
+  subsets: ['latin'],
+  display: 'swap',
+  variable: '--font-mono',
+});
+
+const jetbrainsMono = JetBrains_Mono({
+  subsets: ['latin'],
+  display: 'swap',
+  variable: '--font-jetbrains',
+});
+
+export const metadata: Metadata = {
+  title: 'Kilo Code Prototypes',
+  description: 'Full-page Kilo Code product-flow prototypes for design review.',
+};
+
+export const viewport: Viewport = {
+  themeColor: '#171717',
+  width: 'device-width',
+  initialScale: 1,
+};
+
+export default function RootLayout({ children }: Readonly<{ children: React.ReactNode }>) {
+  return (
+    <html
+      lang="en"
+      className={`${inter.variable} ${mono.variable} ${jetbrainsMono.variable} antialiased`}
+    >
+      <body>{children}</body>
+    </html>
+  );
+}
