+++ apps/prototypes/src/components/shared/Banner.tsx
@@ -1,0 +1,1 @@
+export { Banner } from '@web/components/shared/Banner';
