+++ apps/prototypes/src/components/ui/badge.tsx
@@ -1,0 +1,1 @@
+export * from '@web/components/ui/badge';
