+++ apps/prototypes/src/components/ui/button.tsx
@@ -1,0 +1,1 @@
+export * from '@web/components/ui/button';
