+++ apps/prototypes/src/components/ui/card.tsx
@@ -1,0 +1,1 @@
+export * from '@web/components/ui/card';
