+++ apps/prototypes/src/components/ui/dialog.tsx
@@ -1,0 +1,1 @@
+export * from '@web/components/ui/dialog';
