+++ apps/prototypes/src/components/ui/switch.tsx
@@ -1,0 +1,1 @@
+export * from '@web/components/ui/switch';
