+++ apps/prototypes/src/components/ui/table.tsx
@@ -1,0 +1,1 @@
+export * from '@web/components/ui/table';
