+++ apps/prototypes/src/components/ui/tabs.tsx
@@ -1,0 +1,1 @@
+export * from '@web/components/ui/tabs';
