+++ scripts/lint-all.sh
@@ -22,7 +22,9 @@ workspace_dirs=$(pnpm ls --json -r --depth -1 2>/dev/null | node -e "
 while IFS= read -r pkg; do
   [ -z "$pkg" ] && continue
   if [ -d "$pkg/src" ]; then
-    if [[ "$pkg" == apps/mobile ]]; then
+    if [[ "$pkg" == apps/prototypes ]]; then
+      continue
+    elif [[ "$pkg" == apps/mobile ]]; then
       mobile_lint_dirs+=("$pkg/src")
     else
       lint_dirs+=("$pkg/src")
